Implement i18n
This includes i18n and extract those strings to an english translations files (the default language).

The use of `i18n` rather than `gettext` or other similar things was chosen because:

* it's consistent with anyone wanting to use this gem in a rails app
* It's easier to ship the translatable strings of a gem as yaml files

I've also included `i18n-tasks` as a development dependency so that it is easier to handle missing translations and unused keys.